all test cases done

+++ src/controllers/employee.controller.ts
@@ -1,11 +1,12 @@
 /**
  * Required External Modules and Interfaces
  */
-import { Request, Response, Router } from "express";
+import { Request, Response } from "express";
 import EmployeeService from "../services/employees.services";
 import { EmployeeData, Employee } from "../interfaces/employee.interface";
-import { EnumEmployee } from "../constants/enums";
+import FileService from "../services/FileService.services";
 import { Employees } from "../interfaces/employeeObject.interface";
+import { constants } from "../constants/validationErrors";
 
 const employeeService = new EmployeeService();
 /**
@@ -13,12 +14,31 @@ const employeeService = new EmployeeService();
 */
 export default class EmployeeController {
 
+  private fService: FileService;
+
+  constructor() {
+    this.fService = new FileService()
+  }
+
+  // GET employees by level
+  public getEmployeeLevel = async (req: Request, res: Response) => {
+    try {
+      const emp = await employeeService.getLevel(req.query.type as string);
+      if (emp.length == 0) {
+        res.status(404).json({ "Error": `There are no ${req.query.type}` })
+      }
+      res.status(200).json(emp)
+    } catch (e: any) {
+      return res.status(500).send(e.message)
+    }
+  }
+
   // GET employees
   public getAllEmployees = async (req: Request, res: Response) => {
     try {
-      const employees: Employees= await employeeService.getAll();
-      if(!employees){
-        return res.status(404).json({ "error": "There are no employees in the database." });
+      const employees: Employees = await employeeService.getAll();
+      if (!employees) {
+        return res.status(404).json({ "error": constants.NO_EMPLOYEE });
       }
       return res.status(200).json(employees);
     } catch (e: any) {
@@ -30,9 +50,9 @@ export default class EmployeeController {
   public getEmployee = async (req: Request, res: Response) => {
     const id: number = parseInt(req.params.id, 10);
     try {
-      const employee = await employeeService.find(id);
+      const employee: Employee = await employeeService.find(id);
       if (employee === undefined || employee === null || Object.keys(employee).length == 0) {
-        return res.status(404).json({ "error": `Employee not found with given id ${id}` });
+        return res.status(404).json({ "error": constants.NO_EMPLOYEE_FOUND });
       }
       return res.status(200).json(employee);
     } catch (e: any) {
@@ -40,92 +60,69 @@ export default class EmployeeController {
     }
   }
 
-  // // // // POST employees
+  // POST employees
   public createEmployee = async (req: Request, res: Response) => {
-    const isValidated = await this.validator(req, res);
+    const isValidated = await EmployeeService.validator(req.body, constants.METHOD_CREATE)
     if (!isValidated) {
       try {
         const employee: EmployeeData = req.body;
-        await employeeService.create(employee);
-        return res.status(201).json({"message": "Employee created"});
+        const newEmp = await employeeService.create(employee);
+        return res.status(201).json(newEmp);
       } catch (e: any) {
         return res.status(500).send(e.message);
       }
+    } else {
+      return res.json({ "error": isValidated });
     }
   }
 
-  // // // PUT employees/:id
+  // PUT employees/:id
   public editEmployee = async (req: Request, res: Response) => {
     const id: number = parseInt(req.params.id, 10);
-    const isValidated = await this.validator(req, res);
+    const isValidated = await EmployeeService.validator(req.body, constants.METHOD_EDIT);
     if (!isValidated) {
       try {
-        const employeeUpdate: Employee = req.body;
-        const existingEmployee: Employee = await employeeService.find(id);
-        // console.log(existingEmployee);
-        if (!existingEmployee) {
-          return res.status(404).json({ "error": `Employee not found with given id ${id}` });
+        const employeeUpdate: EmployeeData = req.body;
+        const newEmp = await employeeService.update(id, employeeUpdate);
+        if (!newEmp) {
+          return res.status(404).json({ "error": constants.NO_EMPLOYEE_FOUND});
         }
-        const updatedEmployee = await employeeService.update(id, employeeUpdate);
-        return res.status(200).json(updatedEmployee);
-        // const newEmployee = await employeeService.create(employeeUpdate);
-        // return res.status(201).json(newEmployee);
+        return res.status(200).json(newEmp);
       } catch (e: any) {
         return res.status(500).send(e.message);
       }
+    } else {
+      return res.json({ "error": isValidated });
     }
   }
 
-  // // DELETE employees/:id
+  // DELETE employees/:id
   public deleteEmployee = async (req: Request, res: Response) => {
     try {
       const id: number = parseInt(req.params.id, 10);
-      // const employee = await employeeService.remove(id);
-      await employeeService.remove(id);
-      // if (!employee) {
-        // return res.status(404).json({ "error": `Employee not found with given id ${id}` });
-      // }
-      return res.status(201).json({ "message": `Employee deleted with id ${id}` });
+      if(id === 1 || id === 2 ){
+        return res.json({"error": constants.DELETE_MANAGER_ERROR})
+      }
+      const employee = await employeeService.remove(id);
+      if (!employee) {
+        return res.status(404).json({ "error": constants.NO_EMPLOYEE_FOUND});
+      }
+      return res.status(200).json({ "message": constants.EMPLOYEE_DELETE_ERROR });
     } catch (e: any) {
       return res.status(500).send(e.message);
     }
   }
 
-  public findSubordinates = async(req: Request, res: Response) => {
+  public findSubordinates = async (req: Request, res: Response) => {
     try {
       const id: number = parseInt(req.params.id, 10);
       const employees = await employeeService.getSubordinate(id);
-      if(!employees || Object.keys(employees).length == 0) {
-        return res.status(404).json({ "error": "There are no employees under this user id" });
+      if (Object.keys(employees).length === 0) {
+        return res.status(404).json({ "error": constants.SUBORDINATE_ERROR });
       }
       return res.status(200).json(employees);
-    } catch(e: any){
+    } catch (e: any) {
       return res.status(500).send(e.message)
     }
-  }
-
-  private validator = async(req: Request, res: Response): Promise<boolean> => {
-    const emailReg = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-    const phoneReg =  /^\d{10}$/;
-    const enums = Object.values(EnumEmployee)
-    let validationError = false;
-    
-    if (!req.body.first_name || !req.body.last_name || !req.body.email || !req.body.phone_no || !req.body.level || !req.body.manager) {
-      res.status(400).json({ "Error": "Please pass all required attributes in the body" });
-      validationError = true;
-    }
-    if (!emailReg.test(req.body.email)) {
-      res.status(400).json({ "Error": "email is not valid" });
-      validationError = true;
-    }
-    if (!phoneReg.test(req.body.phone_no)) {
-      res.status(400).json({ "Error": "Phone no is not valid" });
-      validationError = true;
-    }
-    if(!enums.includes(req.body.level)) {
-      res.status(400).json({ "Error": "Level is not valid" });
-      validationError = true;
-    }
-    return validationError;
   }
 }
